pyproject.toml
[project]
name = "DRFTemplate"
version = "1.0.0"
description = "Default template for PDM package"
authors = [
    {name = "Jancel", email = "ja.alvarezgom@gmail.com"},
]
dependencies = [
    "djangorestframework>=3.16.1",
    "pydantic-settings>=2.12.0",
    "Temporis>=1.0.7",
    "psycopg-binary>=3.3.2",
    "gunicorn>=23.0.0",
    "drf-spectacular>=0.29.0",
    "django-sequences>=3.0",
    "django-filter>=25.2",
    "django-cors-headers>=4.9.0",
    # [Celery related packages]
    "celery>=5.6.2",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    # [Redis related packages]
    #    "fakeredis>=2.30.1",
    #    "redis>=5.1.0",
    "djangorestframework-jsonapi>=8.1.0",
    "django-guid>=3.5.2",
    "django-postgres-extra>=2.0.9",
    "google-api-python-client>=2.188.0",
    "google-auth-oauthlib>=1.2.4",
    "boto3>=1.42.30",
    "psycopg>=3.3.2",
    "drf-spectacular-jsonapi>=0.5.2",
    "django-celery-results>=2.6.0",
    "django-activity-stream-with-events>=2.0.1", # Audit package
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest-mock>=3.15.1",
    "pytest-factoryboy>=2.8.1",
    "pytest-django>=4.11.1",
    "ruff>=0.14.13",
    "pytest>=9.0.2",
    "django-schema-graph>=3.1.0",
]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
#celery = "celery -A drf_template worker --concurrency=4 -l INFO -P gevent -Q celery,celery:1,celery:2,celery:3"
cold_start.composite = ["manage.py makemigrations", "manage.py migrate", "seeder --option 99"]
clean_migrations.composite = ['find api/ -path "*/migrations/*.py" -not -name "__init__.py" -delete', 'rm "*.sqlite3"']
rm_sqlite.composite = ['rm "*.sqlite3"']
seeder = "manage.py seeder"
gen_req = "pdm export -o requirements.txt --without-hashes"
lint.composite = ["ruff check --fix", "ruff format"]
test = "pytest"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
minversion = "6.0"
addopts = "-ra -q"
python_files = ["test_*.py", "*_test.py",]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s %(levelname)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 150
indent-width = 4

# Assume Python 3.8
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", ]
ignore = ["F405", "F403", "E402"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

